connect posts to users

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require("express");
 const multer = require("multer");
 
 const Post = require("../models/post");
+const checkAuth= require("../middleware/check-auth");
 
 const router = express.Router();
 
@@ -31,14 +32,15 @@ const storage = multer.diskStorage({
 });
 
 router.post(
-  "",
+  "",checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     const url = req.protocol + "://" + req.get("host");
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -53,41 +55,52 @@ router.post(
 );
 
 router.put(
-  "/:id",
+  "/:id",checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
       const url = req.protocol + "://" + req.get("host");
-      imagePath = url + "/images/" + req.file.filename
+      imagePath = url + "/images/" + req.file.filename;
     }
     const post = new Post({
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if(result.nModified>0) {
+        res.status(200).json({message: "Update successful!"});
+      }else{
+        res.status(401).json({message: "Not"});
+      }
     });
   }
 );
 
 router.get("", (req, res, next) => {
-  const pageSize= +req.query.pageSize;
-  const currentPage= +req.query.page;
-  const postQuery=Post.find();
-  if(pageSize && currentPage){
-    postQuery.skip(pageSize * (currentPage-1))
-      .limit(pageSize);
+  const pageSize = +req.query.pagesize;
+  const currentPage = +req.query.page;
+  const postQuery = Post.find();
+  let fetchedPosts;
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
   }
-  postQuery.find().then(documents => {
-    res.status(200).json({
-      message: "Posts fetched successfully!",
-      posts: documents
+  postQuery
+    .then(documents => {
+      fetchedPosts = documents;
+      return Post.count();
+    })
+    .then(count => {
+      res.status(200).json({
+        message: "Posts fetched successfully!",
+        posts: fetchedPosts,
+        maxPosts: count
+      });
     });
-  });
 });
 
 router.get("/:id", (req, res, next) => {
@@ -100,10 +113,13 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+router.delete("/:id",checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id , creator: req.userData.userId }).then(result => {
+    if(result.n>0) {
+      res.status(200).json({message: "Update successful!"});
+    }else{
+      res.status(401).json({message: "Not"});
+    }
   });
 });
